An initial HALp implementation (#4)
This pull request implements HALp, an add-on for test.frg files, that allows students to get hints from their test suites.

HALp generates hints by first running student tests against a correct implementation. If a test fails:

and is an example, HALp lets users know that the example is invalid.
and is a general test expect, HALp offers no help.
is an assertion, HALp leverages the structure of implications to generate a mutation of the correct implementation that reflects the misconception embodied in the failing assertion. This is then run against an auto-grader, whose results are used to generate feedback.
In doing so, HALp does not focus on every potential mismatch between a correct implementation and the student's tests. Rather, it focuses on those deemed important by the instructor (as reflected in the autograder), and uses those to generate feedback.
---------

Co-authored-by: Siddhartha Prasad <siddhartha.a.prasad@gmail.com>
Co-authored-by: Tim Nelson <tn@cs.brown.edu>