Adding configurable support for different mutation/feedback strategies (#19)
## Comprehensive
This mutation strategy generates a single mutant that reflects the characteristics  of all student test failures. As a result, this mutant reflects the sum (and potentially a more comprehensive view) of student misunderstandings, as defined by the test suite. As a result, aside from a test report, students receive a single piece of feedback that reflects misunderstandings across their test suite. As a result, the effectiveness of suite-level mutation depends on the assumption of internal consistency within student test suites. 

## Per Test

This mutation strategy generates a mutant tailored to each student wheat failing test. Each mutant is then be run against the autograder, resulting in specific feedback associated with each failing test. By considering each test in isolation, feedback can be targeted to each specific test. This may result in more actionable student feedback. This is especially useful if student test suites are not consistent, or if students do not incrementally add tests. However, it runs the risk of information overload, and may present overfitted feedback per test.


---------

Co-authored-by: Siddhartha Prasad <siddharthaprasad@Siddharthas-MacBook-Pro.local>