Release 2.3.0 (#20)
* Changes

* Updating dev guide

* [Staging] Adding support for multiple errors (#14)

### Adding support for multiple errors in a Forge run
This increases the amount / number of conceptual mutations we are able to find from test failures
Users no longer have to comment out tests in order to get additional feedback

### Design Decisions
In the case that multiple hints are available to be served to a user (ie- they fail multiple autograder tests), we pick one at random. This means that they may get different hints for different Toadus Ponens runs

More focused mutations are generated by mean that we give students feedback at the test suite, not test level, since a conceptual mutant may reflect multiple test failures. Still, we show students a list of tests we know to be inconsistent with the problem specification.
---------

Co-authored-by: Siddhartha Prasad <siddharthaprasad@Siddharthas-MacBook-Pro.local>

* Re-enabling logging

* Adding better syntax (#17)

Co-authored-by: Siddhartha Prasad <siddharthaprasad@Siddharthas-MacBook-Pro.local>

* Adding min-supported Forge version enforcement (#16)

* Adding min version check
---------

Co-authored-by: Siddhartha Prasad <siddharthaprasad@Siddharthas-MacBook-Pro.local>

* Adding Thoroughness feedback to Toadus Ponens (#18)

Mutator now can provide thoroughness feedback

---------

Co-authored-by: Siddhartha Prasad <siddharthaprasad@Siddharthas-MacBook-Pro.local>

* Thoroughness better err reporting

* Hotfix -- uncommenting logger

* Adding configurable support for different mutation/feedback strategies (#19)

## Comprehensive
This mutation strategy generates a single mutant that reflects the characteristics  of all student test failures. As a result, this mutant reflects the sum (and potentially a more comprehensive view) of student misunderstandings, as defined by the test suite. As a result, aside from a test report, students receive a single piece of feedback that reflects misunderstandings across their test suite. As a result, the effectiveness of suite-level mutation depends on the assumption of internal consistency within student test suites. 

## Per Test

This mutation strategy generates a mutant tailored to each student wheat failing test. Each mutant is then be run against the autograder, resulting in specific feedback associated with each failing test. By considering each test in isolation, feedback can be targeted to each specific test. This may result in more actionable student feedback. This is especially useful if student test suites are not consistent, or if students do not incrementally add tests. However, it runs the risk of information overload, and may present overfitted feedback per test.


---------

Co-authored-by: Siddhartha Prasad <siddharthaprasad@Siddharthas-MacBook-Pro.local>

* Updating min forge version

* Excluding parameterized examples from mutated feedback (#21)



---------

Co-authored-by: Siddhartha Prasad <siddharthaprasad@Siddharthas-MacBook-Pro.local>

---------

Co-authored-by: Siddhartha Prasad <siddhartha.a.prasad@gmail.com>
Co-authored-by: Siddhartha Prasad <siddharthaprasad@Siddharthas-MacBook-Pro.local>